[server] Refactor StorageReadRequestsHandler and ComputeUtils
