+++ clients/da-vinci-client/src/main/java/com/linkedin/davinci/VersionBackend.java
@@ -7,14 +7,10 @@ import static com.linkedin.venice.ConfigKeys.SERVER_STOP_CONSUMPTION_WAIT_RETRIE
 import com.linkedin.davinci.config.VeniceStoreVersionConfig;
 import com.linkedin.davinci.storage.chunking.AbstractAvroChunkingAdapter;
 import com.linkedin.davinci.store.AbstractStorageEngine;
-import com.linkedin.venice.VeniceConstants;
 import com.linkedin.venice.client.store.streaming.StreamingCallback;
 import com.linkedin.venice.compression.VeniceCompressor;
-import com.linkedin.venice.compute.ComputeOperationUtils;
 import com.linkedin.venice.compute.ComputeRequestWrapper;
-import com.linkedin.venice.compute.ReadComputeOperator;
-import com.linkedin.venice.compute.protocol.request.ComputeOperation;
-import com.linkedin.venice.compute.protocol.request.enums.ComputeOperationType;
+import com.linkedin.venice.compute.ComputeUtils;
 import com.linkedin.venice.exceptions.VeniceException;
 import com.linkedin.venice.meta.IngestionMode;
 import com.linkedin.venice.meta.Store;
@@ -31,7 +27,6 @@ import java.nio.ByteBuffer;
 import java.time.Duration;
 import java.time.Instant;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -42,7 +37,6 @@ import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.apache.avro.Schema;
-import org.apache.avro.generic.GenericData;
 import org.apache.avro.generic.GenericRecord;
 import org.apache.avro.io.BinaryDecoder;
 import org.apache.logging.log4j.LogManager;
@@ -220,7 +214,7 @@ public class VersionBackend {
       BinaryDecoder binaryDecoder,
       ByteBuffer reusableRawValue,
       GenericRecord reusableValueRecord,
-      Map<String, Object> globalContext,
+      Map<String, Object> sharedContext,
       ComputeRequestWrapper computeRequestWrapper,
       Schema computeResultSchema) {
 
@@ -239,12 +233,11 @@ public class VersionBackend {
         storeDeserializerCache,
         compressor.get());
 
-    return getResultOfComputeOperations(
-        computeRequestWrapper.getOperations(),
-        computeRequestWrapper.getValueSchema(),
-        reusableValueRecord,
-        globalContext,
+    return ComputeUtils.computeResult(
         computeRequestWrapper.getComputeRequestVersion(),
+        computeRequestWrapper.getOperations(),
+        sharedContext,
+        reusableValueRecord,
         computeResultSchema);
   }
 
@@ -257,19 +250,18 @@ public class VersionBackend {
       RecordDeserializer<GenericRecord> keyRecordDeserializer,
       GenericRecord reusableValueRecord,
       BinaryDecoder reusableBinaryDecoder,
-      Map<String, Object> globalContext,
+      Map<String, Object> sharedContext,
       Schema computeResultSchema) {
 
     StreamingCallback<GenericRecord, GenericRecord> computingCallback =
         new StreamingCallback<GenericRecord, GenericRecord>() {
           @Override
           public void onRecordReceived(GenericRecord key, GenericRecord value) {
-            GenericRecord computeResult = getResultOfComputeOperations(
-                computeRequestWrapper.getOperations(),
-                computeRequestWrapper.getValueSchema(),
-                value,
-                globalContext,
+            GenericRecord computeResult = ComputeUtils.computeResult(
                 computeRequestWrapper.getComputeRequestVersion(),
+                computeRequestWrapper.getOperations(),
+                sharedContext,
+                value,
                 computeResultSchema);
             callback.onRecordReceived(key, computeResult);
           }
@@ -296,57 +288,6 @@ public class VersionBackend {
         this.storeDeserializerCache,
         this.compressor.get(),
         computingCallback);
-  }
-
-  private GenericRecord getResultOfComputeOperations(
-      List<ComputeOperation> operations,
-      Schema valueSchema,
-      GenericRecord valueRecord,
-      Map<String, Object> globalContext,
-      int computeRequestVersion,
-      Schema computeResultSchema) {
-
-    if (valueRecord == null) {
-      return null;
-    }
-
-    Map<String, String> computationErrorMap = new HashMap<>();
-    GenericRecord resultRecord = new GenericData.Record(computeResultSchema);
-
-    // execute each operation
-    for (ComputeOperation computeOperation: operations) {
-      ReadComputeOperator operator = ComputeOperationType.valueOf(computeOperation).getOperator();
-      String operatorFieldName = operator.getOperatorFieldName(computeOperation);
-      String errorMessage =
-          ComputeOperationUtils.validateNullableFieldAndGetErrorMsg(operator, valueRecord, operatorFieldName)
-              .orElse(null);
-      if (errorMessage != null) {
-        operator.putDefaultResult(resultRecord, operator.getResultFieldName(computeOperation));
-        computationErrorMap.put(operator.getResultFieldName(computeOperation), errorMessage);
-        continue;
-      }
-      operator.compute(
-          computeRequestVersion,
-          computeOperation,
-          valueRecord,
-          resultRecord,
-          computationErrorMap,
-          globalContext);
-    }
-
-    Schema.Field computationErrorMapField =
-        computeResultSchema.getField(VeniceConstants.VENICE_COMPUTATION_ERROR_MAP_FIELD_NAME);
-    if (computationErrorMapField != null && resultRecord.get(computationErrorMapField.pos()) == null) {
-      resultRecord.put(computationErrorMapField.pos(), computationErrorMap);
-    }
-
-    // fill empty fields in result schema
-    for (Schema.Field field: computeResultSchema.getFields()) {
-      if (resultRecord.get(field.pos()) == null) {
-        resultRecord.put(field.pos(), valueRecord.get(field.name()));
-      }
-    }
-    return resultRecord;
   }
 
   public int getPartitionCount() {
